contributed by Natalia:
I gave this BadLocationException second thought and it can simply be ignored. It is caused by concurrent threads modifying document, but this is allowed. We should not even log such exception (cause later it bothers us when we read .logs ;) ). I see that most of places in code where BadLocationException is caught are just empty blocks:
catch (BadLocationException e) { }