Bug 420243 - improvements in ASTUtils methods for use statements
The following three utility methods are improved in performance and
functional correctness:
 - findUseStatementByAlias()
 - findUseStatementByNamespace()
 - getUseStatements()
 
The AST visitors used in the above methods are reworked to traverse only
the nodes where use statements can be legally found and not go deeper in
the AST tree. More details in the JavaDoc.
 
Tests added to the suite.

Bug: 420243
Change-Id: Ia235fca7bc973c8f0a4dc8b7fee3c51cb46a8a7a
Signed-off-by: Kaloyan Raev <kaloyan.r@zend.com>