Bug 436255 - [Library Folders] NPE in UseAsSourceFolderAction
The NPE happens because the explicitly disabled folder is not in the
build path anymore.

This patch refactors the API and implementation in a way that is more
centered around the Resource API instead of the DLTK Model API. This way
folders that are not on the build path can be handled in a better way.

Change-Id: I3be8fe3b4cf6f5bc619318602eb7a08285361584
Signed-off-by: Kaloyan Raev <kaloyan.r@zend.com>